refactor(app): migrate deployment type section [EE-6237]
